feat: only show YouTube sidebar item on demand
