--- resources/assets/js/components/layout/main-wrapper/Sidebar.spec.ts
@@ -1,8 +1,0 @@
-import { expect, it } from 'vitest'
-import UnitTestCase from '@/__tests__/UnitTestCase'
-
-new class extends UnitTestCase {
-  protected test () {
-    it('has already been tested in the integration suite', () => expect('😄').toBeTruthy())
-  }
-}
