+++ resources/assets/js/components/layout/main-wrapper/index.vue
@@ -10,7 +10,7 @@
 <script lang="ts" setup>
 import { defineAsyncComponent } from 'vue'
 
-import SideBar from '@/components/layout/main-wrapper/Sidebar.vue'
+import SideBar from '@/components/layout/main-wrapper/sidebar/Sidebar.vue'
 import MainContent from '@/components/layout/main-wrapper/MainContent.vue'
 import ExtraPanel from '@/components/layout/main-wrapper/ExtraPanel.vue'
