+++ resources/assets/js/components/layout/main-wrapper/sidebar/PlaylistSidebarList.spec.ts
